BerryPicker/requirements.txt
approxeng.input == 2.5
torch
torchvision
pandas
numpy
matplotlib
tqdm
pyyaml
tensorboardX
pyserial
opencv-python
opencv-contrib-python

